Add basic support for parsing commands
Commands are likely to be one of the more complex parts of the grammar.
For this reason support will be implemented incrementally. This basic
support allows the construction of higher level constructs such as
command lists and functions.

The 'word' token is not explicitly defined by the Shell grammar
specification.  It is essentially a context dependant catch-all. The
definition has been restricted to just alpha-numeric to prevent
ambiguity with assignments and other constructs.