Add support for functions
The body of a function can have many forms. The most common one is a brace group
(i.e. '{' command; ... '}'). Since the library is more of a lexer than a parser,
a callback will likely be used to notify the beginning of a brace group, with
brace matching left up to the caller.